samples/microsoft/python/getting-started-agents/hosted-agents/langgraph-multi-agents/requirements.txt
langgraph==1.0.2

azure-ai-agentserver-langgraph==1.0.0b4

pytest==8.4.2
azure-identity==1.25.0
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.8.1
